Working tester getter for position, missing data workflow correctly and go to position function

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/odometry/position/Position.java
@@ -3,7 +3,7 @@ package org.firstinspires.ftc.teamcode.odometry.position;
 public class Position {
     private static double xPosition;
     private static double yPosition;
-    private static int orientation;
+    private static double orientation;
 
     public Position(double xPosition, double yPosition, int orientation) {
         Position.xPosition = xPosition;
@@ -17,11 +17,11 @@ public class Position {
     public static double getYPosition(){
         return yPosition;
     }
-    public static int getOrientation(){
+    public static double getOrientation(){
         return orientation;
     }
 
     public static void setXPosition(double xposition){Position.xPosition = xposition;}
     public static void setYPosition(double yposition){Position.yPosition = yposition;}
-    public static void setOrientation(int orientation){Position.orientation = orientation;}
+    public static void setOrientation(double orientation){Position.orientation = orientation;}
 }
